Mobile: Add support for attaching audio recordings to notes
