+++ .eslintignore
@@ -794,7 +794,10 @@ packages/app-mobile/components/screens/status.js
 packages/app-mobile/components/screens/tags.js
 packages/app-mobile/components/side-menu-content.js
 packages/app-mobile/components/testing/TestProviderStack.js
+packages/app-mobile/components/voiceTyping/AudioRecording.js
+packages/app-mobile/components/voiceTyping/SpeechToText.js
 packages/app-mobile/components/voiceTyping/VoiceTypingDialog.js
+packages/app-mobile/components/voiceTyping/types.js
 packages/app-mobile/gulpfile.js
 packages/app-mobile/index.web.js
 packages/app-mobile/root.js
