+++ packages/app-mobile/components/voiceTyping/types.ts
@@ -1,0 +1,6 @@
+export interface OnFileEvent {
+	uri: string;
+	fileName: string;
+	type: string|undefined;
+}
+export type OnFileSavedCallback = (event: OnFileEvent)=> void;
