+++ packages/app-mobile/web/public/index.html
@@ -10,7 +10,7 @@
 			http-equiv="Content-Security-Policy"
 			content="
 				default-src 'self' ;
-				connect-src 'self' * http://* https://* ;
+				connect-src 'self' * http://* https://* blob: ;
 				style-src 'unsafe-inline' 'self' blob: ;
 				child-src 'self' ;
 				script-src 'self' 'unsafe-eval' 'unsafe-inline' ;
