+++ packages/renderer/MdToHtml/renderMedia.ts
@@ -18,7 +18,7 @@ export interface Options {
 function resourceUrl(resourceFullPath: string): string {
 	if (
 		resourceFullPath.indexOf('http://') === 0 || resourceFullPath.indexOf('https://') === 0 || resourceFullPath.indexOf('joplin-content://') === 0 ||
-		resourceFullPath.indexOf('file://') === 0
+		resourceFullPath.indexOf('file://') === 0 || resourceFullPath.startsWith('blob:null/')
 	) {
 		return resourceFullPath;
 	}
