+++ packages/tools/cspell/dictionary4.txt
@@ -166,3 +166,4 @@ HMAC
 Siri
 cipherdecipher
 pmmmwh
+webm
